# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ResultApp API
  version: 0.1.0
  description: "API for University Result Management (roles: admin, exam_officer, faculty_dean, hod, lecturer, student)"
servers:
  - url: http://localhost:4000/api
paths:
  /auth/register:
    post:
      summary: Register user (admin will create users in production)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
      responses:
        '201':
          description: Created
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: Token
  /users:
    get:
      summary: List users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /results/upload:
    post:
      summary: Lecturer uploads results
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Uploaded
  /results/submit:
    put:
      summary: Lecturer submits results for dept validation
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Submitted
  /results/department-approve:
    put:
      summary: HOD approves department results
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dept approved
  /results/faculty-approve:
    put:
      summary: Faculty Dean recommends/recommends results
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Faculty recommended
  /results/final-approve:
    put:
      summary: Exam Officer final approval & publish
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Published
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Register:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - admin
            - exam_officer
            - faculty_dean
            - hod
            - lecturer
            - student
    Login:
      type: object
      properties:
        email:
          type: string
        password:
          type: string